shows fname
Tamad na ko mag tandog subong
