+++ package.json
@@ -8,6 +8,7 @@
     "start": "next start",
     "lint": "next lint"
   },
+  
   "dependencies": {
     "@emotion/react": "^11.9.0",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
